+++ modules/kor-usecases/src/main/kotlin/com/sefford/kor/usecases/components/NoModule.kt
@@ -1,0 +1,15 @@
+package com.sefford.kor.usecases.components
+
+object NoModule : PerformanceModule {
+
+    override val name: String
+        get() = ""
+
+    override fun start(traceId: String) {
+        // Empty
+    }
+
+    override fun end(traceId: String) {
+        // Empty
+    }
+}